fix: Wayland in split screen mode, split screen preview is triggered. Single touch clicking on the window in the preview box does not trigger split screen.
Wayland in the split screen mode, the split screen preview is not triggered because the corresponding event processing is not implemented, and the window in the preview box is not triggered by single touch.

Log: Wayland in split screen mode, split screen preview is triggered. Single touch clicking on the window in the preview box does not trigger split screen

Bug: https://pms.uniontech.com/zentao/bug-view-111219.html

Change-Id: I7835c0cbff42283f7d0b8f97bf1ace17b3df7b88