fix: In dual-screen mode, exit splitting screen, but splitbar is still displayed
In dual-screen mode, after splitting screen, drag the window away from the splitbar to exit the splitting screen,but splitbar is still displayed.
when a client on screen0, but split screen on screen1, getting screen is also 0.
change to get screen from getConfigGeometry when workspace::updateScreenSplitApp.

Log: change to get screen from getConfigGeometry when workspace::updateScreenSplitApp.
Bug: https://pms.uniontech.com/bug-view-137661.html
Change-Id: Iad41f574af19ed9115564a191857237e82f151a4