Revert "fix: do not transfer focus to invalid toplevel"
This reverts commit 4c97d73e31145bff0312514ef7fb33004101599c.

In V20, this patch has been revert by a patch that includes in
upstream patch, so this patch dose not require. This change will
cause the InternalClient to not be able to get the focus and
corresponding mouse events.
eg:The switcher for plugins with "ALT + TAB" can't select
the client with the mouse.

Log: fix switcher cannot select the client through the mouse
Task: https://pms.uniontech.com/bug-view-141025.html

Signed-off-by: ut004146 <zhangqiyi@uniontech.com>
Change-Id: I44f0fea7ad3f79ddcec48dae3c07abc9940894c9