Fix: fix the problem that the window cannot be maximized
Need to resend the mouse position to the application when
the window changed.
When switching from the maximized to the right split screen and again,
the mouse position does not change, only the press event is reported.
Since the window has changed, the relative coordinates are invalid and
no response is made.

Log: fix the problem that the window cannot be maximized
Task: https://pms.uniontech.com/zentao/bug-view-122483.html

Change-Id: I274c8b951e490fcb8c860d07845f4a568b2c39b1
Signed-off-by: ut004146 <zhangqiyi@uniontech.com>
Signed-off-by: zhang qiyi <zhangqiyi@uniontech.com>