Fix: fix kwin-wayland hang for a while
Just initialize Kwin-Wayland to trigger multitaskview,
this may have a hurry for a while because of the initialization
of QDBusInterface, and caused kwin-wayland hang up. So
add Thread to ensure that Dbus can access normally, not
blocking the main thread of kwin-wayland.

Log: fix kwin-wayland hang for a while
Task: https://pms.uniontech.com/bug-view-137873.html

Change-Id: Ifbe3e93fb05d4222660d579b39a4c936ea447561
Signed-off-by: ut004146 <zhangqiyi@uniontech.com>