fix: fix the problem caused by the conflict of the touchpad gesture
Four-finger DOWN gesture is to showing desktop, and then up is repling
display all window. This function is implemented by dde-daemon.
When the function of showing desktop is not working, the four-finger UP
gesture is to turn on the multi-tasking view.
So avoid conflicting with the touchpad gesture incident of UOS,
turn off Kwin's four-finger gesture event.

Log: fix the problem caused by the conflict of the touchpad gesture
Task: https://pms.uniontech.com/bug-view-141055.html

Change-Id: I510f679109b565bf9e569011b4eebc3f746eea39
Signed-off-by: zhang qiyi <zhangqiyi@uniontech.com>