fix: fix the problem caused by the conflict of the touchpad gesture
Four-finger DOWN gesture is to showing desktop, and then up is repling
display all window. This function is implemented by dde-daemon.
When the function of showing desktop is not working, the four-finger UP
gesture is to turn on the multi-tasking view.
So avoid conflicting with the touchpad gesture incident of UOS,
turn off Kwin's four-finger gesture event.

Log: fix the problem caused by the conflict of the touchpad gesture
Task: https://pms.uniontech.com/bug-view-153177.html
Change-Id: Ib78b948d5da3acd633dda15078682a8f7370f6eb
Signed-off-by: zhang qiyi <zhangqiyi@uniontech.com>