fix: DPMS off seems to blank compositor's back or front buffer
The GLX_EXT_buffer_age allow extension is to expose enough information to
applications about how the driver manages the set of front and back buffers
associated with a given surface to allow applications to re-use the contents
of old frames and minimize how much must be redrawn for the next frame.

If using GLX_EXT_buffer_age on kwin, it's using the repaint area from the
damage areas to do partial refresh in compositor. when dpms off and dpms
on soon(The interval of one second may be repeated), you will see the
screen is flash with the some part of black in some frame, maybe one buffer
of front buffer and back buffer is broke, and kwin only draw a part of the
buffer, so the other part of the buffer is black.

Upstream bug: https://gitlab.freedesktop.org/xorg/xserver/-/issues/1296
Issue: https://github.com/linuxdeepin/developer-center/issues/4252