fix: Avoid crash with on alt+tab(tabbox) view on close it current window
When the window is deleting, if the window is the currentClient
of the TabBox, it will try to reset the currentIndex of the TabBox.
If necessary, the TabBox will update the highlighted window,
and if the kwin compositing is closed at this time, the TabBox
will try to call the restack function. The old currentClient is
placed under the new client. This will cause the currentClient
to be destroy to be re-added to the unconstrained_stacking_order
and crash when accessing the object in the unconstrained_stacking_order

Log: Avoid crash with on alt+tab(tabbox) view on close it current window
Change-Id: Ib9fc6682490c845f7bc21204fd0c4d95e589c2f6