fix: screencasting: on memfd, skip the QImage step
We were using QImage as an intermediary step. GL -> QImage -> spa
buffer. While it abstracted things out neatly and eventually helped with
debugging, it was unnecessary and woudl present some handicaps, such as
the lack of a QImage::Format_BGRA.
So we just it out to download right into the buffer.

5ac5251674d910e3402885144dc6ce9479a960eb

Log: screencasting: on memfd, skip the QImage step
Change-Id: I87bbb272bac657e43a4903eaf620d8c9412baf3b