fix: deepin-screen-recorder recording GIF, kwin_wayland crashes
when remote_access_manager remove rbuf, it will check if BufferHandle
is not nullptr. But in current code we use built-in pointer local
variable as if's condition. This is not affect the original pointer of
the external scope.

To fix this, we use QPointer instead of introduce high-level pointers,
QPointer can remember its null status when it be copyed.

Log: deepin-screen-recorder recording GIF, kwin_wayland crashes
Bug: https://pms.uniontech.com/bug-view-281413.html
Bug: https://pms.uniontech.com/bug-view-283017.html
Change-Id: Ica945f7dddc3f006136b00ef5393a75d55c74ae5