fix: delay done event after all changes are emitted
Event notifications in the code about the kde_output_device_v2 protocol
are sent as soon as a change occurs, not strictly as described in IDL.
In some cases (such as plugging a new screen), client B may get the wrong
screen for a short period of time because kwin sends the default initialized
screen status to client B immediately before receiving the new screen
configuration set by client A. If client B were something like a desktop-wallpaper,
this would cause a certain amount of flickering. On the other hand, it will
 increase the amount of data traffic, and a lot of messages will be invalid.
The improved scheme imitates wl_output protocol and adds a doneTimer to realize
the debounce and batching of event notification.

Bug: https://pms.uniontech.com/bug-view-275061.html
Bug: https://pms.uniontech.com/bug-view-263971.html

Log: delay done event after all changes are emitted
Change-Id: I5b7add789d28716fbe17a206e2c970a19659355c