feat: make dock can use systemResize
X11 clients can use _NET_WM_MOVERESIZE to let the WM
take charge of the resize operation, so as to have a
smooth resize effect. However, the implementation in
Kwin prohibits this behavior of special window by default.
This will cause the dock to be unable to call this mechanism,
resulting in a relatively poor drag and drop experience.

So in this PR, I released the _NET_WM_MOVERESIZE request for
dock type windows. At the same time, when the window type is
Dock, moveResize is set to a Unrestricted state to remove the
influence of the exclusive area.

log: as title