fix: crash when output added and removed
Presumably, if the compositor drawing is activated and the event is
generated after the output is unref to 0, then the wild pointer will be
accessed during the frame skip.

In postPaint, it is necessary to access painted_screen, which is updated
in prePaint, but there is no such call in the frame skip.

PMS: https://pms.uniontech.com/bug-view-302251.html
Log:
Change-Id: Ie8ab2c686894fe36aad16f424758c8f728c79503