fix: adapt offscreenquickview to Qt6 changes
Qt6 documentation states:

"Note: When using this window in combination with QQuickRenderControl,
the result of this function is an empty image, unless the software
backend is in use. "

We support a mode of using openGL rendering but then exporting as pixmap
images for Aurorae. In this case we should simply grab data from our
FBO.

see: https://github.com/KDE/kwin/commit/ced95f536cf4119e2c3a7579636fe27290d94486