fix: crash in OffscreenQuickView::update when rhi is null
This is merely a temporary workaround. On machines with NVIDIA dedicated GPUs that trigger a multi-display setup, a crash may occur due to a null RHI (Rendering Hardware Interface) pointer. Skipping this update can prevent the crash, though the side effect is that users may notice the loss of one frame update.